% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_deforestation.R
\name{deforestation}
\alias{deforestation}
\title{Deforestation Function}
\usage{
deforestation(
  natural_forest_t1,
  natural_forest_t2,
  plantation_t1,
  plantation_t2,
  fire_risk_score,
  nat_weight = 0.8,
  plan_weight = 0.3
)
}
\arguments{
\item{natural_forest_t1}{natural forest area at earlier time point (ha)}

\item{natural_forest_t2}{natural forest area at later time point (ha)}

\item{plantation_t1}{timber plantation area at earlier time point (ha)}

\item{plantation_t2}{timber plantation area at later time point (ha)}

\item{fire_risk_score}{regional fire risk score from wildfire_risk() Module 3 (unitless)}

\item{nat_weight}{ecological weight for natural forest loss (0-1), default 0.8}

\item{plan_weight}{ecological weight for plantation loss (0-1), default 0.3}
}
\value{
deforestation factor (ha weighted by forest type; higher = greater loss)
}
\description{
Estimates deforestation as net change in forest area, distinguishing
natural forest from timber plantation
}
