% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_habitat_frag.R
\name{habitat_frag}
\alias{habitat_frag}
\title{Habitat Fragmentation Function}
\usage{
habitat_frag(natural_forest_t2, total_landscape_area, defor_factor)
}
\arguments{
\item{natural_forest_t2}{remaining natural forest area at later time point (ha)}

\item{total_landscape_area}{total area of the region being assessed (ha)}

\item{defor_factor}{deforestation factor from deforestation() Module 4 (unitless)}
}
\value{
fragmentation score (unitless; higher = more fragmented landscape)
}
\description{
Estimates landscape fragmentation from deforestation and remaining forest area.
Connectivity is derived internally from remaining natural forest rather than
supplied as an external input, since it is not an output of any upstream module.
}
