% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_human_geog.R
\name{human_geog}
\alias{human_geog}
\title{Human Geography Function}
\usage{
human_geog(pop_t1, pop_t2)
}
\arguments{
\item{pop_t1}{total population at earlier census period (persons)}

\item{pop_t2}{total population at later census period (persons)}
}
\value{
net population change between census periods (persons, positive = growth)
}
\description{
Computes net population change from census data at two time points
}
