% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_land_use.R
\name{land_use}
\alias{land_use}
\title{Land Use Function}
\usage{
land_use(forest_t1, forest_t2, ag_t1, ag_t2, urban_t1, urban_t2)
}
\arguments{
\item{forest_t1}{forest area at earlier time point (ha)}

\item{forest_t2}{forest area at later time point (ha)}

\item{ag_t1}{agricultural area at earlier time point (ha)}

\item{ag_t2}{agricultural area at later time point (ha)}

\item{urban_t1}{urban area at earlier time point (ha)}

\item{urban_t2}{urban area at later time point (ha)}
}
\value{
land use change coefficient (ha lost from forest; positive = forest loss)
}
\description{
Computes a land use change coefficient from change in land cover area
}
