% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_wildfire_risk.R
\name{wildfire_risk}
\alias{wildfire_risk}
\title{Wildfire Risk Function}
\usage{
wildfire_risk(
  rainfall,
  hectares_burned,
  avg_temp,
  net_pop_change,
  landuse_coeff
)
}
\arguments{
\item{rainfall}{total annual rainfall (mm)}

\item{hectares_burned}{hectares affected by fire in prior year (ha)}

\item{avg_temp}{average annual temperature (degrees C)}

\item{net_pop_change}{net population change from human_geog() Module 1 (persons)}

\item{landuse_coeff}{land use change coefficient from land_use() Module 2 (ha)}
}
\value{
regional fire risk score (unitless index, higher = greater risk)
}
\description{
Estimates regional fire risk from climate, disturbance, and land use inputs
}
